Fix XSS vulnerability in sanitizer.py
The YUI editor returns user input in proper HTML so e.g. "<script>"
becomes &lt;script%gt;.  This is fine and should be a good start in
preventing XSS.  But for some reason the sanitizer was parsing the
user input HTML with
soup = BeautifulSoup(html, convertEntities=BeautifulSoup.XHTML_ENTITIES)
which would turn &lt; back to "<".  This content was not being parsed for
HTML content resulting in a wide open platform for XSS.